- Created RollBackManager - Added NetworkManager - Established local connection to send testing inputs

+++ Assets/_Scripts/Data/SimulationConfig.cs
@@ -7,7 +7,7 @@ namespace Data
     {
         public int TargetFPS;
         public int MinutesPerMatch;
-        public int InputDelay;
+        public ushort InputDelay;
         public int BufferSize;
     }
 }
